sql: minor improvements around estimated row count
This commit makes it so that we use the ceiling of the estimated row
count in scans and aggregation. It also fixes recently-introduced bug
that could cause integer overflow in the agg alloc code.

Release note: None

+++ pkg/sql/distsql_spec_exec_factory.go
@@ -286,7 +286,7 @@ func (e *distSQLSpecExecFactory) ConstructScan(
 			spans:             spans,
 			reverse:           params.Reverse,
 			parallelize:       params.Parallelize,
-			estimatedRowCount: uint64(params.EstimatedRowCount),
+			estimatedRowCount: params.EstimatedRowCount,
 			reqOrdering:       ReqOrdering(reqOrdering),
 		},
 	)
